Change detection & async pipes for play / pause
Change detection was not being triggered automatically on play & pause
of the html audio element, probably because the audio element is
running outside of Angular's zone.  As a result, in order to implement
reactive logic based on pause status, I was previously explicitly
subscribing to the subject and explictly calling change detection within
the relevant components.  It worked, but it's a very messy way to do
things.

I fixed this by moving the audio element's play / pause BehaviorSubject
logic to the PlayerService and using applicationRef.tick to explictly
tell Angular to detect changes throughout the app on play / pause status
change.  This has the benefit of allowing me to remove the subscription
& change detection logic from my components and use an async pipe
instead.