Now Playing info refactor
This is a big refactor in which I
-Updated the structure of the NowPlaying metadata model to store both
station info and stream info in their original form
-Renamed the now-playing metadata model to stream-info to reduce
confusion
-Updated the NowPlaying logic such that any time the station or stream
info is updated, it's checked against the old value for equality and a
new value is emitted only if there's a change.
-Updated PlayerService to always emit a new object reference, rather
than a mutation of the previous object.
-Refactored the "Loading" and "Error" text logic such that those states
are represented on the NowPlaying model and components can decide how
they want to display these states accordingly (rather than hardcoding it
in the song title).
-Implemented Now Playing template binding via async pipe, rather than
component property subscription